# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Swimlane Studio API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/tickets/{id}": {
      "delete": {
        "tags": [
          "Tickets"
        ],
        "summary": "Delete Ticket",
        "operationId": "delete_ticket_api_tickets__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/project": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Project",
        "operationId": "project_api_project_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Projects"
        ],
        "summary": "Update Project",
        "operationId": "update_project_api_project_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "Tickets",
        "operationId": "tickets_api_tickets_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default",
              "title": "Project Id"
            }
          },
          {
            "name": "deleted",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Deleted"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ticket"
                  },
                  "title": "Response Tickets Api Tickets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tickets"
        ],
        "summary": "Create",
        "operationId": "create_api_tickets_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Draft"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{ticket_id}": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "Get",
        "operationId": "get_api_tickets__ticket_id__get",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Tickets"
        ],
        "summary": "Update",
        "operationId": "update_api_tickets__ticket_id__patch",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Changes"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{ticket_id}/run": {
      "post": {
        "tags": [
          "Runs"
        ],
        "summary": "Run",
        "operationId": "run_api_tickets__ticket_id__run_post",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChangedFile": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "additions": {
            "type": "integer",
            "title": "Additions"
          },
          "deletions": {
            "type": "integer",
            "title": "Deletions"
          },
          "diff": {
            "type": "string",
            "title": "Diff"
          }
        },
        "type": "object",
        "required": [
          "path",
          "additions",
          "deletions",
          "diff"
        ],
        "title": "ChangedFile"
      },
      "Changes": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "demo",
                  "codex",
                  "claude"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "images": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 6
              },
              {
                "type": "null"
              }
            ],
            "title": "Images"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          },
          "permission": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "read-only",
                  "workspace-write"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Permission"
          },
          "allow_tests": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allow Tests"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          }
        },
        "type": "object",
        "title": "Changes"
      },
      "Draft": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "prompt": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Prompt"
          },
          "provider": {
            "type": "string",
            "enum": [
              "demo",
              "codex",
              "claude"
            ],
            "title": "Provider",
            "default": "demo"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 6,
            "title": "Images"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id",
            "default": ""
          },
          "permission": {
            "type": "string",
            "enum": [
              "read-only",
              "workspace-write"
            ],
            "title": "Permission",
            "default": "read-only"
          },
          "allow_tests": {
            "type": "boolean",
            "title": "Allow Tests",
            "default": false
          },
          "group_id": {
            "type": "string",
            "maxLength": 80,
            "title": "Group Id",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title",
          "prompt"
        ],
        "title": "Draft"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Project": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "default": "My first project"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "screenshot": {
            "type": "string",
            "title": "Screenshot",
            "default": ""
          }
        },
        "type": "object",
        "title": "Project"
      },
...
      "Ticket": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "prompt": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Prompt"
          },
          "provider": {
            "type": "string",
            "enum": [
              "demo",
              "codex",
              "claude"
            ],
            "title": "Provider",
            "default": "demo"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 6,
            "title": "Images"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id",
            "default": ""
          },
          "permission": {
            "type": "string",
            "enum": [
              "read-only",
              "workspace-write"
            ],
            "title": "Permission",
            "default": "read-only"
          },
          "allow_tests": {
            "type": "boolean",
            "title": "Allow Tests",
            "default": false
          },
          "group_id": {
            "type": "string",
            "maxLength": 80,
            "title": "Group Id",
            "default": ""
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id",
            "default": "default"
          },
          "status": {
            "type": "string",
            "enum": [
              "todo",
              "queued",
              "running",
              "done",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "default": "todo"
          },
          "response": {
            "type": "string",
            "title": "Response",
            "default": ""
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/ChangedFile"
            },
            "type": "array",
            "title": "Files"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "queued_at": {
            "type": "string",
            "title": "Queued At",
            "default": ""
          },
          "latest_run_id": {
            "type": "string",
            "title": "Latest Run Id",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title",
          "prompt",
          "id",
          "created_at"
        ],
        "title": "Ticket"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}